+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next
 
 - **[Breaking change]** Require Node `13.2`.
+- **[Breaking change]** Replace `readString` and `readCString` with checked `readUtf8` and `readNulTerminatedString`.
 - **[Fix]** Create `CHANGELOG.md`.
 - **[Fix]** Update dependencies.